Make precommit hook os-dependant
This commit creates seperate precommit hooks for darwin and linux.
precommitStatisticsHook-darwin.sh is a copy of the original
precommitStatisticsHook.sh. In the linux version, linux-incompatible
commands are commented out. README is updated accordingly.

Makefiles under common, newton and noisy are modified to copy the
correct precommit hook based on $OSTYPE defined in config.local.