Make SMT2 backend check for div-by-zero error
This patch makes the SMT2 backend to create an assertion in the
form of `(not (= 0 b))` for any expression in the form of `(/ a b)`

This is required as SMT LIB does not require divide by zero to be
an error. In fact, solvers are free to define the value of a/0.