Use newtonParseQuantityExpression in newtonParseExponentialExpression
At the momemnt, newtonParseExponentialExpression uses
newtonParseNumericExpression, which generates a malformmed quantity
expression tree. This patch changes the use of
newtonParseExponentialExpression to newtonParseQuantityExpression, which
is known to generate a correct quantity epxression tree.

The dimension check that follows is commented out both to avoid
unexpected behaviour and because it should be part of the dimension
check pass instead.