Add generate Linear Kalman Filter prediction eqs
Estimator Synthesis:
Add made assumptions. These will change.
Add function to detect whether a passed expression is linear, that is,
it is able to be reduced to quantityTerms that do not contain
exponentations and more than a given number of identifiers. This can be
more strict and needs to.
Remove printing of identifier in matched expressions by substituting
with 1.
Include matrix library. Subject to change.
Add if-else for linear and nonlinear case of models.
Add generate filterPredict() kalman filter matrix equations. Not tested.

Note that noises (here Q) is missing due to the distribution grammar not having been
incorporated yet.

Other:
Add integerValue to parsed integerConst nodes.
Add case to irPassCNodeToStr() for above nodes.

Addresses #463.