Add findExpressionIdentifiers to common-irHelpers
Function is essentially moved to common-irHelpers to be of use in other
parts of the codebase.

The function provides an interface to find the unique identifiers used in a
`quantityExpression`. As it is now, it returns a list of IrNodes with the shallow copies
of the original IrNodes.

Can be of use when creating argument lists for a specific expression.

Addresses #463.