Ensure allocated object is initialized
I observed a bug where I would get Segmentation Faults in
irPassEstimatorSynthesisCreateConstraintList(), in the case of
kNewtonIrNodeType_Pconstraint, where the code is looking for the last
element of the list in listHead.

The code made a comparison against NULL to check if the element is the
last element. Sometimes the iter-next field contained garbage which lead
to the SIGSEV signal. This is because malloc() does not make any
guarantees about the contents of the allocated space in the heap.

This commit changes the malloc() calls to equivalent calloc() calls to
ensure the space is set to zero on allocation. Another possible solution
is to manually set `iter->next = NULL` after the relevant malloc().

Addresses #542.