Emit extra params in arguments
The backend was only emitting state variables as parameters to the
process predict functions, while the relation matrix did have
information on all variables. Ensure that the relation matrix is
correctly iterated to emit all relations in emitted function arguments.

A similar problem is present in the measurement update emitted code but
it seems more complicated to fix, because the measurement invariant also
relates measurement variables with the state variables.

Addresses #542.