Add := operator for read and arithmetic sensor commands
Add the assign-define or definition operator (`:=`) as part of the
Newton lexical tokens. This operator helps distinguish between
identifier definitions and identifier usages in places where it was
previously ambiguous (like in the read sensor command).

The operator can (but is currently not) also be used in invariants for
defining shorthants for repeating expressions.

Addresses #559.