Added instrumentation for the runtime function `newtonInsert`.
When encountering a variable definition, we insert a call to malloc
in order to allocate an array-state for the variable. Then, we add
a call to `newtonInsert` where we send the array as an argument for
the runtime to record and we get back a unique identifier. This will
be used as a future reference for our variable's dimension array.

Addresses #586.